Move js and img to src dir
