+++ _sass/_general.scss
@@ -35,7 +35,7 @@ h1 {
 }
 
 .ev-header {
-  background: #2a2a2b url("img/ev-bg.jpg") center top no-repeat;
+  background: #2a2a2b url("./src/img/ev-bg.jpg") center top no-repeat;
   background-size: cover;
   text-align: center;
   padding: 90px 0 30px;
