+++ index.html
@@ -9,6 +9,6 @@ layout: compress
     {% include partials/gtm.html %}
     {% include partials/header.html %}
     <div id="react-app"></div>
-    <script src="./js/bundle.js"></script>
+    <script src="./src/js/bundle.js"></script>
   </body>
 </html>
