+++ src/js/components/Event/Event.js
