+++ src/js/components/Event/EventApp.js
