+++ src/js/components/Event/EventEmptyResult.js
