+++ src/js/components/Event/EventFilter.js
