+++ src/js/components/Event/EventList.js
