+++ src/js/components/Event/EventTypeRadio.js
