+++ test/components/Event/EventApp.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventApp from '../../../js/components/Event/EventApp';
+import EventApp from '../../../src/js/components/Event/EventApp';
 
 it('should work', () => {
   const wrapper = shallow(<EventApp />);
