+++ test/components/Event/EventEmptyResult.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventEmptyResult from '../../../js/components/Event/EventEmptyResult';
+import EventEmptyResult from '../../../src/js/components/Event/EventEmptyResult';
 
 it('should work', () => {
   const wrapper = shallow(<EventEmptyResult />);
