+++ test/components/Event/EventFilter.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventFilter from '../../../js/components/Event/EventFilter';
+import EventFilter from '../../../src/js/components/Event/EventFilter';
 
 it('should work', () => {
   const wrapper = shallow(<EventFilter />);
