+++ test/components/Event/EventList.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventList from '../../../js/components/Event/EventList';
+import EventList from '../../../src/js/components/Event/EventList';
 
 it('should work', () => {
   const wrapper = shallow(<EventList />);
