+++ test/components/Event/EventSearch.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventSearch from '../../../js/components/Event/EventSearch';
+import EventSearch from '../../../src/js/components/Event/EventSearch';
 
 it('should work', () => {
   const wrapper = shallow(<EventSearch />);
