+++ test/components/Event/EventTypeRadio.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventTypeRadio from '../../../js/components/Event/EventTypeRadio';
+import EventTypeRadio from '../../../src/js/components/Event/EventTypeRadio';
 
 it('should work', () => {
   const wrapper = shallow(<EventTypeRadio />);
