+++ test/components/Event/EventYears.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EventYears from '../../../js/components/Event/EventYears';
+import EventYears from '../../../src/js/components/Event/EventYears';
 
 it('should work', () => {
   const wrapper = shallow(<EventYears />);
